switching over to chrono - nanosecond presicion + beginning of shade cache + Vsync capability

+++ include/core/globals.hpp
@@ -9,6 +9,6 @@ GLFWwindow* mainWindow;
 Shader* mainShader;
 Shader* lightShader;
 
-float deltaTime;
+double deltaTime;
 
 #endif // GLOBAL_PROJECT_VARIABLE_HEADER
